+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Use pwix:logger universal logger, thus bumping minor candidate version number
     - Align border radius on bootstrap measures
     - Extend link validity to have 'From ... to ...'
+    - Define new 'validity-period-created', 'validity-period-left-merged', 'validity-period-right-merged' and 'validity-period-removed' events
 
 ### 1.3.0